Initialize i as double

+++ src/provagit/ProvaGit.java
@@ -3,7 +3,7 @@ package provagit;
 public class ProvaGit {
 
     public static void main(String[] args) {
-        int i=0;
+        double i=0;
         
         while (true) {
             System.out.println("Ehi, ciao! "+i);
